requirements-service.txt
# Browser Automation Service Dependencies
# Core service framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0

# Browser automation (existing dependencies)
browser-use==0.1.33
playwright==1.45.0

# LLM integrations
langchain-google-genai==1.0.8
openai==1.3.5

# Async and utilities
asyncio-mqtt==0.11.1
python-multipart==0.0.6
python-dotenv==1.0.0

# Monitoring and logging
structlog==23.2.0

# Prompt management
PyYAML==6.0.1